+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [1.9.2]
 ### Added user functionality
+- Add RTF export to DossierExport (#703)
 - Allow mailability to be set at multiple levels (#603)
+- Allow save & resume from LimeSurvey using GT (#702)
 - Allow some (limited) option for repeating parts of tracks (#697)
 - Auto mail job should also be able to send different kind of messages (#636)
 - Displaying answers depending on role (#23)